requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-ninja==1.4.1
djangorestframework==3.16.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
kombu==5.5.2
prompt_toolkit==3.0.50
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
